review type resolver
