+++ PostgresOutbox/Events/EventsAppender.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using Npgsql;
 using PostgresOutbox.Serialization;
-using PostgresOutbox.Subscriptions.Replication;
 
 namespace PostgresOutbox.Events;
